fix: links & styles
