+++ src/app/components/AccountMenu/index.tsx
@@ -14,6 +14,7 @@ import Avatar from "~/app/components/Avatar";
 import MenuDivider from "~/app/components/Menu/MenuDivider";
 import { useAccount } from "~/app/context/AccountContext";
 import { useAccounts } from "~/app/context/AccountsContext";
+import { isAlbyAccount } from "~/app/utils";
 import msg from "~/common/lib/msg";
 import utils from "~/common/lib/utils";
 
@@ -75,45 +76,44 @@ function AccountMenu({ showOptions = true }: Props) {
   }
 
   return (
-    <div className="relative pl-2 w-72 flex border border-gray-200 rounded-md dark:bg-surface-12dp">
-      <div className="flex items-center">
-        <Avatar size={24} name={authAccount?.id || ""} />
-      </div>
-
-      <div
-        className={`flex-auto mx-2 py-3 overflow-hidden ${
-          !title && !balancesDecorated ? "w-28" : ""
-        }`}
-      >
-        <p
-          title={title || ""}
-          className="text-xs text-gray-700 dark:text-neutral-400 text-ellipsis overflow-hidden whitespace-nowrap"
-        >
-          {title || <Skeleton />}
-        </p>
-      </div>
-
+    <div className="relative pl-2 flex justify-end w-72">
       <Menu as="div">
-        <Menu.Button className="h-full px-2 rounded-r-md hover:bg-gray-200 dark:hover:bg-white/10 transition-colors duration-200">
-          <CaretDownIcon className="h-4 w-4 dark:text-white" />
-          <span className="sr-only">{t("screen_reader")}</span>
+        <Menu.Button className="h-full px-2 rounded-md hover:bg-gray-200 dark:hover:bg-white/10 transition-colors duration-200">
+          <div className="flex items-center">
+            <Avatar size={24} name={authAccount?.id || ""} />
+            <div
+              className={`flex-auto mx-2 py-3 overflow-hidden ${
+                !title ? "w-28" : ""
+              }`}
+            >
+              <p
+                title={title || ""}
+                className="text-sm font-medium text-gray-700 dark:text-neutral-400 text-ellipsis overflow-hidden whitespace-nowrap"
+              >
+                {title || <Skeleton />}
+              </p>
+            </div>
+            <CaretDownIcon className="h-4 w-4 dark:text-white" />
+            <span className="sr-only">{t("screen_reader")}</span>
+          </div>
         </Menu.Button>
-
-        <Menu.List position="left" fullWidth>
+        <Menu.List position="right" fullWidth>
           <Menu.Item>
             <div
-              className={`flex-auto mx-5 py-2 overflow-hidden ${
+              className={`flex-auto px-4 py-2 overflow-hidden ${
                 !title && !balancesDecorated ? "w-28" : ""
               }`}
             >
-              <span className="text-xs">Balance</span>
+              <span className="text-xs text-gray-500 dark:text-neutral-300">
+                Balance
+              </span>
               {balancesDecorated.accountBalance ? (
                 <p className="flex justify-between">
                   <span className="text-md dark:text-white">
                     {balancesDecorated.accountBalance}
                   </span>
                   {!!balancesDecorated.fiatBalance && (
-                    <span className="text-xs text-gray-600 dark:text-neutral-400 ml-2">
+                    <span className="text-gray-500 dark:text-neutral-300">
                       ~{balancesDecorated.fiatBalance}
                     </span>
                   )}
@@ -125,25 +125,25 @@ function AccountMenu({ showOptions = true }: Props) {
           </Menu.Item>
           <Menu.ItemButton
             onClick={() => {
-              console.log("OK");
+              navigate(`/accounts/${authAccount?.id}`);
             }}
           >
             <WalletIcon className="h-5 w-5 mr-2 text-gray-700 dark:text-neutral-300" />
             {t("options.account.account_settings")}
           </Menu.ItemButton>
-          <Menu.ItemButton
-            onClick={() => {
-              console.log("OK");
-            }}
-          >
-            <GlobeIcon className="h-5 w-5 mr-2 text-gray-700 dark:text-neutral-300" />
-            {t("options.account.go_to_web_wallet")}
-          </Menu.ItemButton>
-
+          {isAlbyAccount(authAccount?.alias) && (
+            <Menu.ItemButton
+              onClick={() => {
+                window.open(`https://getalby.com/login`, "_blank");
+              }}
+            >
+              <GlobeIcon className="h-5 w-5 mr-2 text-gray-700 dark:text-neutral-300" />
+              {t("options.account.go_to_web_wallet")}
+            </Menu.ItemButton>
+          )}
           <MenuDivider />
 
           <Menu.Subheader>{t("title")}</Menu.Subheader>
-
           {Object.keys(accounts).map((accountId) => {
             if (accountId === authAccount?.id) {
               return;
