+++ src/app/screens/Accounts/index.tsx
@@ -43,9 +43,8 @@ function AccountsScreen() {
                 >
                   <td className="px-6 py-4 whitespace-nowrap">
                     <div className="flex items-center">
-                      <div className="w-12 h-12 flex justify-center items-center">
-                        <Avatar name={account.id} size={48} />
-                      </div>
+                      <Avatar name={account.id} size={48} />
+
                       <div className="ml-4">
                         <h3 className="font-bold text-gray-900 dark:text-white break-all whitespace-normal max-w-xs md:max-w-lg xl:max-w-2xl">
                           {account.name}
